sendgrid 'add' email now works& moved to add PUT req instead of own folder, update installs, changeurls and auth redirects to work on NOW
