+++ .gitignore
@@ -1,6 +1,7 @@
 .DS_Store
 .git
 .env
+deploy.sh
 
 # dependencies
 node_modules
